refactor and streamline music route, remove route slots
