+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "kalynbeach-xyz",
-  "version": "0.5.1",
+  "version": "0.6.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
