--- src/app/(music)/music/@enter/enter-scene.tsx
@@ -1,12 +1,0 @@
-import Canvas from "@/components/canvas";
-import EnterSphere from "./enter-sphere";
-
-export default function EnterScene() {
-  return (
-    <Canvas>
-      <ambientLight />
-      <pointLight position={[10, 10, 10]} />
-      <EnterSphere />
-    </Canvas>
-  );
-}
