+++ src/app/(music)/music/loading.tsx
@@ -1,0 +1,8 @@
+export default function Loading() {
+  console.log(`[music] Loading...`);
+  return (
+    <div className="h-full w-full flex items-center justify-center">
+      Loading...
+    </div>
+  );
+}
