+++ src/components/theme-provider.tsx
@@ -1,4 +1,4 @@
-"use client"
+"use client";
 
 import * as React from "react";
 import { ThemeProvider as NextThemesProvider } from "next-themes";
