+++ src/components/wave-player-container.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { StackProvider, WavePlayerStack } from "@kalynbeach/wave-player";
-import { tracks } from "./content";
+import { tracks } from "@/app/(music)/music/content";
 
 export default function WavePlayerContainer() {
   return (
